LXD Default logger
The following moves the lxdclient logger to the container.
